+++ src/Fields/Relations/HasOne.php
@@ -1,6 +1,6 @@
 <?php
 /*
- * Copyright 2021 Cloud Creativity Limited
+ * Copyright 2022 Cloud Creativity Limited
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -23,7 +23,7 @@ use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\HasOne as EloquentHasOne;
 use Illuminate\Database\Eloquent\Relations\MorphOne as EloquentMorphOne;
 use LaravelJsonApi\Eloquent\Contracts\FillableToOne;
-use LaravelJsonApi\Eloquent\Fields\Concerns\ReadOnly;
+use LaravelJsonApi\Eloquent\Fields\Concerns\IsReadOnly;
 use LogicException;
 
 class HasOne extends ToOne implements FillableToOne
@@ -33,7 +33,7 @@ class HasOne extends ToOne implements FillableToOne
     private const DELETE_DETACHED_MODEL = 1;
     private const FORCE_DELETE_DETACHED_MODEL = 2;
 
-    use ReadOnly;
+    use IsReadOnly;
 
     /**
      * Flag for how to detach a model from the relationship.