+++ src/QueryBuilder/EagerLoading/EagerLoadIterator.php
@@ -1,6 +1,6 @@
 <?php
 /*
- * Copyright 2021 Cloud Creativity Limited
+ * Copyright 2022 Cloud Creativity Limited
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -24,6 +24,7 @@ use Illuminate\Support\Str;
 use IteratorAggregate;
 use LaravelJsonApi\Core\Query\IncludePaths;
 use LaravelJsonApi\Eloquent\Schema;
+use Traversable;
 
 /**
  * Class EagerLoadIterator
@@ -95,7 +96,7 @@ class EagerLoadIterator implements IteratorAggregate
     /**
      * @inheritDoc
      */
-    public function getIterator()
+    public function getIterator(): Traversable
     {
         /**
          * We always need to yield the default paths on the base schema.