+++ src/QueryMorphToMany.php
@@ -1,6 +1,6 @@
 <?php
 /*
- * Copyright 2021 Cloud Creativity Limited
+ * Copyright 2022 Cloud Creativity Limited
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -109,7 +109,7 @@ class QueryMorphToMany implements QueryManyBuilder, IteratorAggregate
     /**
      * @return Generator
      */
-    public function getIterator()
+    public function getIterator(): Generator
     {
         foreach ($this->relation as $relation) {
             $query = $this->toQuery($relation);