+++ tests/lib/Acceptance/SoftDeleteTest.php
@@ -1,6 +1,6 @@
 <?php
 /*
- * Copyright 2021 Cloud Creativity Limited
+ * Copyright 2022 Cloud Creativity Limited
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -140,7 +140,7 @@ class SoftDeleteTest extends TestCase
 
         $this->schema->repository()->delete((string) $post->getRouteKey());
 
-        $this->assertDeleted($post);
+        $this->assertModelMissing($post);
         $this->assertTrue($forceDeleted);
     }